destination_service/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
PyJWT==2.10.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
Werkzeug==3.1.3
